[feat] Reshuffle how cards work a lot and tidy up
