+++ client/src/app/cards/breaks.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useRef, useState } from "react"
-import { setHoliday, setHost, submitClaim } from "./api"
+import { setHoliday, setHost, submitClaim } from "../api"
 import {
     User,
     CookieBreak,
@@ -13,13 +13,13 @@ import {
     formatAsPrice,
     getDatetimeText,
     getShortDate,
-} from "./structs"
+} from "../structs"
 import {
     BreakControlIcons,
     BreakStatusIcons,
     DeleteBreakIcon,
     SmallIcon,
-} from "./icons"
+} from "../icons"
 import {
     CardAction,
     CardSelector,
@@ -31,8 +31,7 @@ import {
     SmallInfoCard,
 } from "./cards"
 import { ClaimBreakCost } from "./claims"
-
-export type SetState<T> = React.Dispatch<React.SetStateAction<T>>
+import { SetState } from "../page"
 
 export const TickCrossInputBox = (props: {
     onClickClose: (text: string) => void
@@ -198,14 +197,14 @@ const BreakDate = (props: { cb: CookieBreak }) => {
     )
 }
 
-const BreakDetails = (props: {
+export const BreakDetails = (props: {
     user: User | undefined
     cb: CookieBreak
     updateBreaks: UpdateBreaksFn
     setCardLoading: (loading: boolean) => void
 }) => {
     let detailsStyle =
-        "flex flex-col justify-around tablet:flex-row " + "items-center flex-1"
+        "flex flex-col justify-around tablet:flex-row items-center flex-1"
     return (
         <div className={detailsStyle}>
             <BreakDate cb={props.cb} />
@@ -271,9 +270,9 @@ const BreakCard = (props: {
     </>
 )
 
-const getCardColour = (cb: CookieBreak) =>
+export const getCardColour = (cb: CookieBreak) =>
     cb.holiday ? "bg-gray-200" : "bg-white"
-const getSelectedColour = (_: CookieBreak) => "bg-gray-100"
+export const getSelectedColour = (_: CookieBreak) => "bg-gray-100"
 const getHoverColour = (_: CookieBreak) => "hover:bg-gray-50"
 
 export const BreakCards = (props: {
@@ -315,7 +314,6 @@ export const BreakCards = (props: {
             elements={props.breaks}
             getCardColour={getCardColour}
             getCardContent={getCardContent}
-            buttons={props.buttons}
         />
     )
 }
@@ -332,161 +330,6 @@ const BreakReimbursedDate = (props: { cb: CookieBreak }) => {
             icon="reimburse"
             alt="Coin"
             content={content}
-        />
-    )
-}
-
-const AwaitingClaimCard = (props: {
-    user: User | undefined
-    cb: CookieBreak
-    setLoading: (loading: boolean) => void
-    updateBreaks: UpdateBreaksFn
-}) => {
-    const cardStyles =
-        "flex flex-col desktop:flex-row w-full justify-center " +
-        "items-center justify-between p-2 desktop:p-0"
-    return (
-        <div className={cardStyles}>
-            <div className="flex-1">
-                <BreakDetails
-                    user={props.user}
-                    cb={props.cb}
-                    updateBreaks={props.updateBreaks}
-                    setCardLoading={props.setLoading}
-                />
-            </div>
-            <div className="flex flex-row justify-end w-5/12">
-                <BreakReimbursedDate cb={props.cb} />
-                <div className="w-1/4 my-auto">
-                    <ClaimBreakCost cb={props.cb} />
-                </div>
-                <DeleteBreakIcon
-                    user={props.user}
-                    cb={props.cb}
-                    updateBreaks={props.updateBreaks}
-                    setCardLoading={props.setLoading}
-                />
-            </div>
-        </div>
-    )
-}
-
-const AwaitingClaimInfoCards = (props: { cb: CookieBreak }) => {
-    let content = (
-        <>
-            {!props.cb.reimbursed
-                ? ""
-                : `Reimbursed on ${getDatetimeText(props.cb.reimbursed)}`}
-        </>
-    )
-    return (
-        <>
-            <SmallInfoCard
-                width="w-7/12"
-                icon="reimburse"
-                alt="Coin"
-                content={content}
-            />
-        </>
-    )
-}
-
-const AwaitingClaimCardExpanded = (props: {
-    user: User | undefined
-    cb: CookieBreak
-    setLoading: (loading: boolean) => void
-    updateBreaks: UpdateBreaksFn
-}) => {
-    const cardStyles = "flex w-full flex-col"
-    const expansionStyles =
-        "m-2 p-2 border-t-2 flex flex-row flex-wrap justify-center"
-    return (
-        <div className={cardStyles}>
-            <AwaitingClaimCard
-                user={props.user}
-                cb={props.cb}
-                setLoading={props.setLoading}
-                updateBreaks={props.updateBreaks}
-            />
-            <div className={expansionStyles}>
-                <AwaitingClaimInfoCards cb={props.cb} />
-            </div>
-        </div>
-    )
-}
-
-export const AwaitingClaimCards = (props: {
-    user: User | undefined
-    breaks: CookieBreak[]
-    updateBreaks: UpdateBreaksFn
-    updateClaims: UpdateClaimsFn
-    isLoadingBreaks: boolean
-}) => {
-    const [breaksToClaim, setBreaksToClaim] = useState<CookieBreak[]>()
-    useEffect(() => {
-        setBreaksToClaim(getBreaksToClaim(props.breaks))
-    }, [props.breaks])
-    const getCardContent = (
-        cb: CookieBreak,
-        setLoading: (loading: boolean) => void
-    ) => (
-        <AwaitingClaimCard
-            user={props.user}
-            cb={cb}
-            setLoading={setLoading}
-            updateBreaks={props.updateBreaks}
-        />
-    )
-    const getCardContentExpanded = (
-        cb: CookieBreak,
-        setLoading: (loading: boolean) => void
-    ) => (
-        <AwaitingClaimCardExpanded
-            user={props.user}
-            cb={cb}
-            setLoading={setLoading}
-            updateBreaks={props.updateBreaks}
-        />
-    )
-    const makeClaim = (
-        cbs: CookieBreak[],
-        setLoadingCards: (cbs: CookieBreak[], loading: boolean) => void
-    ) => {
-        if (props.user) {
-            submitClaim(
-                props.user,
-                cbs,
-                props.updateBreaks,
-                props.updateClaims,
-                (b) => setLoadingCards(cbs, b)
-            )
-        }
-    }
-    const claimButton = {
-        buttonName: "Make claim",
-        submitSelection: makeClaim,
-        flavourText: (cbs: CookieBreak[]) => {
-            let cost = cbs.reduce(
-                (acc, cur) => acc + (cur.cost ? cur.cost : 0),
-                0
-            )
-            return formatAsPrice(cost)
-        },
-    }
-    const cardsAction: SelectableCardsProps<CookieBreak> = {
-        type: CardAction.SELECT,
-        getSelectedColour: getSelectedColour,
-        getHoverColour: getHoverColour,
-        buttons: [claimButton],
-    }
-    return (
-        <Cards<CookieBreak>
-            title="Awaiting claim"
-            getCardColour={getCardColour}
-            getCardContent={getCardContent}
-            cardsAction={cardsAction}
-            isLoading={props.isLoadingBreaks}
-            elements={breaksToClaim}
         />
     )
 }
